uk-pmrv-app-web/src/app/permit-application/approaches/approaches-summary/approaches-summary.component.ts
import { ChangeDetectionStrategy, Component } from '@angular/core';
import { Router } from '@angular/router';

import { PermitApplicationStore } from '../../store/permit-application.store';

@Component({
  selector: 'app-approaches-summary',
  templateUrl: './approaches-summary.component.html',
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class ApproachesSummaryComponent {
  notification = this.router.getCurrentNavigation()?.extras.state?.notification;

  constructor(readonly store: PermitApplicationStore, private readonly router: Router) {}
}
